template instantiations for semantic actions
