+++ src/stan/lang/grammars/bare_type_grammar.hpp
@@ -3,6 +3,7 @@
 
 #include <boost/spirit/include/qi.hpp>
 #include <stan/lang/ast.hpp>
+#include <stan/lang/grammars/semantic_actions.hpp>
 #include <stan/lang/grammars/whitespace_grammar.hpp>
 
 #include <string>
