+++ src/stan/lang/grammars/statement_2_grammar.hpp
@@ -2,8 +2,9 @@
 #define STAN_LANG_GRAMMARS_STATEMENT_2_GRAMMAR_HPP
 
 #include <stan/lang/ast.hpp>
-#include <stan/lang/grammars/whitespace_grammar.hpp>
 #include <stan/lang/grammars/expression_grammar.hpp>
+#include <stan/lang/grammars/semantic_actions.hpp>
+#include <stan/lang/grammars/whitespace_grammar.hpp>
 #include <boost/spirit/include/qi.hpp>
 #include <sstream>
