+++ src/stan/lang/grammars/statement_grammar.hpp
@@ -4,9 +4,10 @@
 #include <stan/lang/ast.hpp>
 #include <stan/lang/grammars/expression_grammar.hpp>
 #include <stan/lang/grammars/indexes_grammar.hpp>
+#include <stan/lang/grammars/semantic_actions.hpp>
 #include <stan/lang/grammars/statement_2_grammar.hpp>
-#include <stan/lang/grammars/var_decls_grammar.hpp>
 #include <stan/lang/grammars/whitespace_grammar.hpp>
+#include <stan/lang/grammars/var_decls_grammar.hpp>
 #include <boost/spirit/include/qi.hpp>
 #include <sstream>
 #include <string>
